No outdent at eof (#22435)
Fixes #22332 

The reference does not mention `<eof>`. `<outdent>` insertion does not
require a particular next token, though some next tokens affect it (that
is, leading infix or tokens that close an indentation region). It does
require a "first token on the next line", for which we may take `<eof>`
as the lack of a token.

Of course, ordinary error messages say `eof`.

The same text with an opening brace is unchanged:
```
5 |
  | ^
  | '}' expected, but eof found
```
[Cherry-picked 81e057a8d4bcde2d8cee628b8b02f90fbfb05954]