Fix stack overflow errors when generating opaque type proxies (#22479)
Before the regressive PR, we would check via the generated opaqueProxies
list whether one was already generated. In that PR, we tried allowing
generating proxies for rhs of currently generated proxy. Since we have
to add the generated proxy to opaqueProxies only after that step, this
could cause infinite recursion (and adding the proxies earlier could
cause another infinite loop).

To fix that, we add another collection for termrefs which we already
visited this way, but which is not used in the `mapOpaques` function.