Allow observing an indent after conditional
Normally do not infer NEWLINE within parens,
but special case old syntax for conditionals,
so that it can observe indented syntax.
The mechanism is to inject an Indented region
when parsing a parenthesized condition which is
within an InParens region, such as an arg list.
The effect is not to advance past EOL after `(true)`.

[Cherry-picked 4ba89e9e8eb027b2bd11503e3621fa5e62e7bad0]