Fix issue with static `this` references erroring in quoted code
Previously, inherited methods, even if accessed via static objects,
were not able to be used in quotations, unless explicitly pointed to
with a non-`this` prefix. This was due to the fact that during the
cross-stage safety check, first the method itself was checked for if it
was static (which for the inherited method, it was not), and if not,
the prefix was checked further, erroring on any `this` tree found along
the way.

This was fixed by allowing `this` trees if they point to static objects.
This way not only is the initial issue fixed, but also we are able to
freely reference static methods with `this`, like '{this.objectMethod}
(whereas previously, only '{Object.objectMethod} or '{objectMethod}
were allowed, despite them all pointing to the same static method).

[Cherry-picked 71ddfb5f628b3d5c9b02d187f8fbbf6c03147283]