Open existential scopes only for `(x: T) => U` functions
Parametric functions `A -> B` or `A => B` do not open an existential scope
anymore. Hence, a function of the form
```scala
  def f(x: A => B^)
```
is capture polymorphic in both the function effect and the function result effect.
On the other hand,
```scala
  def f(x: (y: A) => B^)
```
would be a function that can be passed only arguments that return a fresh result on each call.

[Cherry-picked 30971b574332dfe735d2c6886e30dee7f6b74633]