Change rules for tracked refinments and tighten intersections
Intersections used the heuristic mightAccountFor instead of the precise accountsFor. Thsi
can lead to a loss of precision and (if unchecked afterwards) also soundness. The fix caused
some tests to fail, which involved tracked parameters.

We now deal with tracked parameters in the same way as parameters that carry a @refineOverride annotation.

[Cherry-picked 23cb1f3622f954e583eb8bc0b052ee7a9a880e27]