Fix #22794: Emit the definition of `Arrays.newArray` even though it's a primitive.
Until 3.6.x, we used the user-space definition of `Arrays.newArray`
in Scala.js. We changed it to a primitve in
2852168e8ddac8aa318635a73c79f64282aa39d1, to match the behavior of
the JVM backend.

However, that had the side effect of not emitting its definition
at all, like other primitives. For backward binary compat reasons,
we do need that definition.

We specifically make an exception for that method, which is now
emitted, even though it is otherwise treated as a primitive.