Mitigate change in status of scala.caps
`scala.caps`` was an object until 3.6, it is a package from 3.7. Without special handling
this would cause a TypeError to be thrown if a build has several versions of the
Scala standard library on the classpath. This was the case for 29 projects in the open CB.
These projects should be updated. But until that's the case we issue a warning instead
of a hard failure.