Fix issue with certain synthetics missing in compiletime.typechecks
Previously, instead of using the PostTyper phase directly, a typed tree
Transformer was used. However, it did not transform the ClassInfo
values using PostTyper.transformInfo, which lead to RefChecks not
finding certain synthesized methods, throwing an incorrect error about
a missing implementation. This is fixed by running the full suite of
PostTyper transforms. A minor refactor is included, to limit
the repeated CompilationUnit construction and error checking.

[Cherry-picked 8f48a3befb3b23ec6cb9fbfca1bc3e148c0d893b]