Revert TastyFormat changes
Instead use a hack to convert the Into flag into the $into annotation for pickling.
This means we can merge the scheme without waiting for the next minor version.

When the scheme is stabilized we should revert this commit again to get the cleaner
treatment where the Into fag maps to the INTO tag.