Upgrade to Scala.js 1.19.0.
With minimal changes to preserve the status quo. These are forward
ports of compiler, library and build changes from the following
upstream commits:

* https://github.com/scala-js/scala-js/commit/b38201c0cf6f99fc146d0ee90de989b23d6e9c68
  Drop support for non-strict floats.
* https://github.com/scala-js/scala-js/commit/82910a086fe6133e20242b74938a61c0b0084643
  Move well-known names to a new object `WellKnownNames`.
* https://github.com/scala-js/scala-js/commit/53dc4fe54568acfeb1fd31e89030e98a538bd9c6
  Introduce `NewLambda` to synthesize instances of SAM types.
* https://github.com/scala-js/scala-js/commit/9481522bcd4251bda3ccee0cc57863826b7738f1
  Rewrite old IR with `AnonFunctionN` references to use `NewLambda`.

From the `NewLambda` commit, we only adapt the way we compile
Scala function lambdas. We do not use `NewLambda` for arbitrary
SAM types yet. This is left for a future, independent commit, as
it is not required for correctness.

[Cherry-picked 12ca4588636fe97048902d4c6eb5b9390549f7f7][modified]