improvement: Rework IndexedContext to reuse the previously calculated scopes (#22898)
It turns out the work being done in IndexedContext was already done in
Completions, but better, since it doesn't try to read files as the
separate logic does.

There is still some improvement to be done to not calculate it twice,
but in order to keep this PR as simple as possible I will skip that for
now.
[Cherry-picked 9d90ff5f5f7525785799c6a5844a25930ab4cb92][modified]