Don't elide capture sets of capability classes in most cases
When printing types deriving from capability, do print their capture sets even if these
are implied. This avoids confusing errors where a capability classes has two different implied
Fresh instances which are both elided so the types look the same.

Exception: When printing a singleton type (x: T) we do elide in T, since the usually the only
part that matters is the `x`. But we don't do the elision under -Ycc-verbose or -explain.