Prevent fallback to normal function when comparing dependent functions
A dependent function is represented as a refinement type with a normal applied
function type as parent. When capture checking two dependent function types, the
parents should be ignored. But there was a case where this did not happen:
If the left hand side is a sinletontype with a capturing type over a dependent function
type, and the original method type comparisons fail, then the parent type of the
left hand type was tried as a fallback. This commit adds a test that prevents
the fallback.