Allow caps in result types of functions to be mapped to reach capabilities
Example: If we have `val f: A => C^` then `f` has type `A => C^{f*}`. Before it had
that type only if `A` was pure. On the other hand, result caps are not replaced
anymore. We used to have for `val f: (x: A) => C^` with `A` pure that `f: (x: A) => C^{f*}`.
That's no longer the case. In this case, `f` now has type `f: (x: A) => C^` with the `^`
understood to be existentially bound.

This aligns the implementation with the Capless paper.