Warn trivial recursion with module prefix (#23278)
Fixes #23277 

First, loosen the test for whether to warn for `module.recur()`. The
`this` assignment has a RHS that is not pure (because of module init)
but `isStable` (and not significant with respect to the warning).

Second, avoid emitting such a `this` assignment if the prefix is the
enclosing module. The existing code checks for a prefix that is
`Enclosing.this`.

Other "aliasing" prefixes are not considered, such as a method with
result `this.type`, since they could execute arbitrary intended effects.