Turn assert in RetainingType into test
We used to have the case where a RetainingType needs to be tested after
a CapturingType. It would be better to reverse that order: If a RetainingType
is possible test this first, and then test a CapturingType. If we just test for
CapturingType we should crash if it's a RetainingType. That way we can express
expectations that a retaining type was transformed.

This commit does the first half: RetainingType unapply now can handle a CapturingType
(and respond with a None). We still need to do the second half, that a CapturingType
unapply does not handle a RetainingType.