Add an -Yimplicit-to-given flag for rewrites to easily test changes in the ecosystem (#22580)
The flag replaces every implicit keyword with a given or using, aside from
a few exceptions.
The added flag should be used in conjunction with `--rewrite`, adding
rewrite patches to the source code. This will omit implicit classes and
old-style implicit conversions, where a simple rewrite to `given` or
`using` might not be available. `using` keywords are also added to
converted previously-implicit-now-using parameter applications, but only
if both the method definition and application are in the same
compilation run (in other cases they would have to be added manually).