Fix issue with macro results having unexpected source switches
Problem seems to have been derived from the fact that we would only
remap the source of the outermost tree inserted in a Hole, instead of
going through the whole tree inserted there. This meant that when
inserting `Block` with `Imports` in the contents, the Block node would
have a new source assigned (causing a source change encoded in the
TASTy), and all contents would still have the previous source assigned
(causing more source changes encoded).

Now we reassign sources to nodes recursively, meaning that only
the outermost node has a source change assigned in TASTy.