Drop PruneErasedDefs
The motivation for the phase was always unclear to me. No tests broke after the change.

Piggybacked on the phase was the unrelated "anonymous class in inline method" warning.
I moved it in changed form to FirstTransform. It is more efficient now since it does
not need an auxiliary tree traversal.