Fix isPureExpr test
We assumed all type applications were pure, which is claerly false.

E.g.

    def fn[T] = println("hi")
    fn[Int]

This triggered a problem in coverage tests which now assumes a use of an erased
function is not erased (probably because it now shows up in the coverage info).
The problem was fixed by mapping the erased defs to inline defs.

I found out the problem with coverage: If an argument is not a pure expression, it gets lifted
into a prefix val. And the prefix val is not erased even if the parameter is erased. So that way
things escape an erased context. I believe this is a fundamental problem that would also affect
lifting due to other reasons (e.g. eta expansion or handling default arguments). We either
have to fix the lifting to mainain erased status, or restrict erased arguments to pure expressions
that don't need lifting.