Fix regressions in asSeenFrom introduced in 3.7
The body of isLegalPrefix used to read:

    pre.isStable || !ctx.phase.isTyper

but was changed to drop the second condition in #21954 (originally included in
3.6.4-RC1, reverted from 3.6.4 final but back in 3.7). This has led to a
number of regressions, the last ones discussed in #23423.

To make the testcases added in #21594 pass, this PR proposes a less drastic
change: relax isLegalPrefix as before, unless we're doing an implicit search.
This should dramatically reduce the possibility for regressions.

Fixes #23423.
Fixes #22676.

[Cherry-picked 2e4bc0abb8a374d78fe1aac383f3208b5efb5eb1][modified]