-Yprofile-trace properly report macro splicing source
Previously when we had some code like:

def userCode: Unit =
  ...
  val x = myExpensiveMacro()
  ...

The profiler reported the time spent executing the macro as `«method userCode»`
which is misleading. This commit changes it to be reported as `«method
myExpensiveMacro»`. This is particularly useful in bottom-up profiling to find
which macros we spend the most amount of time in.