More systematic handling of invalid prefixes
It looks like there are many situations where an illegal higher-kinded type in an argument for
a value type parameter causes illegal TermRefs and TypeRefs to be constructed, leading to an
assertion error. We now turn the assertion error into a specialized exception which eventually
leads to a TypeError being thrown.

The problem is we cannot detect illegal kinds in arguments early enough to prevent these
situations. We do detect them later, but the damage can already be done before that.

[Cherry-picked 4a959b1181f4a17d9722fbe0fda010da435e9cf4]