Re-use SkolemType for arguments
There are at least two instances where we might skolemize the type of an argument.
One is in safeSubstParams when we compute a result type of a dependent method type
application. The other is in constrainIfDependentParamRef where we constrain type
variables representing dependent parameters.

This used to give several skolem types that ended up in a common OrType. We now
store computed skolem types in attachments of argument types and re-use them if
possible.