Refinement: Make scheme work also if arguments are duplicated
The new scheme works also if argument trees are not unique. Instead of storing
skolem types directly in attachment of arguments, we store a map from arguments
to their skolems in the enclosing application node. But we check then that the
same argument tree does not appear several times in an application. If it does
appear several times, we don't generate the map.

Furthermore, we are also safe if the whole application tree is used several times,
since for each call we generate new skolems and store them in the map. So different
calls get different skolems.