Use result of lambda type of implicit in CheckUnused (#23497)
Fixes #23494 

When inspecting unused implicit parameters, the check skips parameters
which are "marker traits". It tests for any members of the type (or its
upper bound) which are not "universal members". This commit uses the
`resultType` to avoid an error
```
invalid new prefix
```
while computing members where the type is a `LambdaType`.

A future improvement would be not to request `allMembers`, since the
check only needs to find one that is not universal.

The necessitating change was
https://github.com/scala/scala3/commit/2e4bc0abb8a374d78fe1aac383f3208b5efb5eb1.
[Cherry-picked a5e15ab7d7b271272c9e1ac1d2837425ec592214]