Do not generate certain interface mixin forwarders if not necessary
This affects indirectly extended java interfaces, bringing
the behavior closer to how scala 2 handles it. After this commit, the
methods from those interfaces will not be generated, if they are
unnecessary (e.g., where there is no conflict between parents when
calling that method) even if `-Xmixin-force-forwarders` is set to true.
Those previously added mixin forwarders caused meant that since specific
parents were referenced in those methods, they had to be added to the
`interfaces` array in the generated classfile (without being added to
the signature, which caused an issue with getInterfaces and
getGenericInterfaces not being equal).