Use unsigned arithmetics in Range, instead of Longs.
Previously, `Range` used a number of intermediate operations on
`Long`s to avoid overflow. We can streamline a lot of code by
using unsigned `Int` arithmetics. In particular, there is only 1
division in the initialization path, instead of 3.

Although the fields have not changed, the content of
`numRangeElements` is more strict for overfull ranges. This means
that deserializing an overfull range from a previous version would
not be safe. This is why we bump the SerialVersionUID.

This commit upstreams
https://github.com/scala-js/scala-js/commit/d9722185beae7b66772b3bf9f91894a781613fe1
from Scala.js.