Improve closure typing
If the closure has an expected function type with a fully defined result type,
take the internalized result type as the local return type of the closure.

This has the effect that some conformance tests are now done with Fresh instead Result
caps. This means a now can widen a local reference to a result cap, since the comparison
is done between the local reference and the internalized FreshCap. Previously this failed
since we compared a local cap with result cap, and result caps only subtype other result caps.

It also propagates types more aggressively into closure bodies, which sometimes reduces the
error span and improves the error message.