Make `caps.Capability` non-experimental (#23507)
Closes #22745.

Make `scala.caps.Capability` non-experimental.

The `Capability` trait can be used as a marker trait even outside of
capture-checking.
Only when capture checking is enabled, the trait carries a special
meaning, marking each instance a top-level capability.

Other definitions inside the `scala.caps` package (the package is
already non-experimental) stays experimental
and can only be used with `-experimental` or with capture-checking
enabled (with an import or with the feature flag).