Scala.js: Handle `@JSName` annots with constant-folded arguments.
Previously, if an `@JSName` annotation had an argument that was
not a literal, but a reference to a constant expression (such as
a `final val`), it would not be constant-folded in the generated
Scala.js IR.

This produced worse code than necessary. For Wasm, it was
particularly bad, as the names must then be evaluated on the Wasm
side instead of being pushed to the custom JS helpers.