By default don't include span captures in the underlying capset
What is the underlying capture set of a function? CC says the capture set
on the first arrow but we experimented with the capture of the span of the
function instead. This however causes a lot of complications if we want to push
it to a logical conclusion. So we now make it configurable and by default
span capture set is not used.