Scala.js: Optimize varargs in a platform-dependent friendly way.
We now recognize the shapes of varargs, deconstruct them, and emit
instead calls to Scala.js-specific runtime methods. Those methods
choose at link-time the best implementation for the target platform.

The changes in `genActualArgs` revealed that we cannot trust
`Symbol.isPrivate`. It uses `lastDenot.flagsUNSAFE` under the hood,
which returns absurd results if `lastDenot` happened to be accessed
within an `atPhase(...)` call. Therefore, we change those calls to
`sym.is(Private)` instead. `Symbol.is` always uses an up-to-date
denotation.

Another consequence of the optimization is that the string
representation of varargs seqs is altered in Scala.js. This is
consistent with the Scala 2 behavior. We generalize the check file
lookup of Vulpix to allow platform-dependent check files. We use
that mechanism to introduce Scala.js-specific versions of the
affected tests.

This commit is essentially a forward port of the Scala.js commit
https://github.com/scala-js/scala-js/commit/24750cfba3f3f2fbfb091fa29fabda6cb0cac165
except that, in dotc, previously we did not do the optimization at
all.