Scala.js: Add support for the JS operator `**`.
This is a forward port of the compiler changes in the Scala.js commit
https://github.com/scala-js/scala-js/commit/5bfa254920eaf4470bb19bc40951a5860c551486

It is now possible to use `@JSOperator` on a method named `**` in
a JS type. In that case, it will translate to the JavaScript `**`
operator. This is only valid when emitting ES 2016 or later, as it
is not valid ES 2015 code.

Without an explicit `@JSOperator` annotation, such a method still
defaults to being a method call, to preserve backward
compatibility.