Warn if empty template name has trailing colon
This is to guard the edge case `object X_:` where the user
may or may not have intended colon syntax. The next line does
not tell us, since it may be indented yet not nested. Therefore,
any empty template with a suspicious name will warn. Non-empty
templates are given a pass even if written `object X_: :`.