When unpickling, do not incorrectly set PureInterface flags
Previously, the behavior for setting `NoInits` and `PureInterface` flags
for class symbols in `TreeInfo.defKind` (when compiling from a source)
and in `TreeUnpickler` (when reading Tasty) was inconsistent.
This was problematic, as the `PureInterface` flag dictated whether outer
accessors were generated in `ExplicitOuter` phase (in
the `needsOuterIfReferenced` method), so in the issue minimization,
the inherited trait would have the outer accessors defined, and
the inheriting object would then not implement them. This lead to
runtime `MethodNotFound` errors.
The problem was specifically with lazy vals, which were treated as if
they did not have an rhs.