Distinguish between certain capset vars with same owner
A capset var that is defined in the result type of a val or def cannot accept any
capabilities defined on the RHS. By contrast, a capset defined in the RHS (but not
in local definitions of the RHS) can accept other capabilities defined in that RHS.
Both of these capsets have the same owner, so we need an additional piece of information
to perform level checking correctly.

Fixes #24039.