SepCheck: Revise definition of peaks
New definition does not go upwards from hidden to hiding, but does
go into hidden sets. Except for a variant `exposedPeaks` which does
not do that.

This is currently very rough. There's a switch to go from old to new behavior
based on source version.

Todo: Polish and then drop the switch if we are convinced verything works as expected.