disable again non-bootstrapped tests
Now that we should have been patching Range.scala too, the code generated tries to call a function from the stdlib (`Object StrictSeqOps.appendedAll$(...)`) that will not exist because in the non-bootstrapped generated library, the generated method is `scala.caps.Pure ...`. This is not an issue in the bootstrapped version since we erase scala.caps.Pure correctly.