fix: Fix incremental compilation of macros
**Problem**
sbt-deps phase runs after the Typer, but prior to the Inliner.
This means that the Zinc won't be able to track the symbolic
dependency from the use site to the generated code,
unless the code exists as a form of quotation.
This often results in under-compilation of the macro use sites.

**Solution**

1. Move the sbt-deps phase after the Inliner, which expands the macros
2. Traverse the Inlined tree, and treat it like a normal code